+++ internal/driver/db_operations.go
@@ -21,6 +21,9 @@ const (
 	createIndex = iota
 )
 
+// queries is a map that stores the SQL queries for different database types.
+// The keys of the outer map represent the database types (e.g., SQLite, MySQL),
+// and the values are inner maps that store the specific queries for each database type.
 var queries = map[string]map[int]string{
 	SQLite: {
 		checkTable:  "SELECT * FROM %s WHERE 1=2",
@@ -44,27 +47,31 @@ var queries = map[string]map[int]string{
 	},
 }
 
+// Column represents a database column.
 type ColumnType map[string]string
 type Column struct {
-	Name       string
-	Type       ColumnType
-	Properties string
-	Index      bool
+	Name       string     // Name of the column.
+	Type       ColumnType // Type of the column.
+	Properties string     // Additional properties of the column.
+	Index      bool       // Indicates if the column has an index.
 }
 
+// Table represents a database table.
 type Table struct {
-	Name           string
-	PrimaryKeyName string
-	PrimaryKeyType ColumnType
-	Columns        []Column
+	Name           string     // Name of the table.
+	PrimaryKeyName string     // Name of the primary key column.
+	PrimaryKeyType ColumnType // Type of the primary key column.
+	Columns        []Column   // List of columns in the table.
 }
 
+// View represents a database view.
 type SQLQuery map[string]string
 type View struct {
-	Name  string
-	Query SQLQuery
+	Name  string   // Name of the view.
+	Query SQLQuery // SQL query used to create the view.
 }
 
+// NewView creates a new View with the given name and SQLQuery.
 func NewView(name string, query SQLQuery) *View {
 	return &View{
 		Name:  name,
@@ -72,6 +79,7 @@ func NewView(name string, query SQLQuery) *View {
 	}
 }
 
+// NewTable creates a new Table instance with the specified name, primary key name, primary key type, and columns.
 func NewTable(name string, primaryKeyName string, primaryKeyType ColumnType, columns []Column) *Table {
 	return &Table{
 		Name:           name,
@@ -81,6 +89,8 @@ func NewTable(name string, primaryKeyName string, primaryKeyType ColumnType, col
 	}
 }
 
+// initTable initializes the table by checking if it exists and creating it if necessary.
+// It also ensures that all required columns exist in the table.
 func (t *Table) initTable(db *sql.DB, engine string) error {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
@@ -101,6 +111,7 @@ func (t *Table) initTable(db *sql.DB, engine string) error {
 	return nil
 }
 
+// isExists checks if the table exists in the database.
 func (t *Table) isExists(ctx context.Context, db *sql.DB, engine string) bool {
 	query := fmt.Sprintf(queries[engine][checkTable], t.Name)
 	_, err := db.ExecContext(ctx, query)
@@ -108,6 +119,7 @@ func (t *Table) isExists(ctx context.Context, db *sql.DB, engine string) bool {
 	return err == nil
 }
 
+// createTable creates the table in the database.
 func (t *Table) createTable(ctx context.Context, db *sql.DB, engine string) error {
 	query := fmt.Sprintf(queries[engine][createTable], t.Name, t.PrimaryKeyName, t.PrimaryKeyType[engine])
 	_, err := db.ExecContext(ctx, query)
@@ -118,6 +130,10 @@ func (t *Table) createTable(ctx context.Context, db *sql.DB, engine string) erro
 	return nil
 }
 
+// ensureColumnsExist ensures that all columns defined in the Table struct exist in the database table.
+// It queries the database to retrieve the existing columns and compares them with the columns defined in the Table struct.
+// If a column is missing, it adds the column to the table using an ALTER TABLE query.
+// If the column has an index, it also ensures that the index exists in the database.
 func (t *Table) ensureColumnsExist(ctx context.Context, db *sql.DB, engine string) error {
 	query := fmt.Sprintf(queries[engine][getColumns], t.Name)
 	rows, err := db.QueryContext(ctx, query)
@@ -163,6 +179,7 @@ func (t *Table) ensureColumnsExist(ctx context.Context, db *sql.DB, engine strin
 	return nil
 }
 
+// ensureIndexExists checks if an index exists in the specified table for the given column.
 func (t *Table) ensureIndexExists(ctx context.Context, db *sql.DB, engine string, column Column) error {
 	query := fmt.Sprintf(queries[engine][checkIndex], t.Name)
 	rows, err := db.QueryContext(ctx, query)
@@ -194,6 +211,7 @@ func (t *Table) ensureIndexExists(ctx context.Context, db *sql.DB, engine string
 	return err
 }
 
+// createIndex creates an index for the specified column in the table.
 func (t *Table) createIndex(ctx context.Context, db *sql.DB, engine string, column Column) error {
 	indexName := fmt.Sprintf("%s_%s", t.Name, column.Name)
 	query := fmt.Sprintf(queries[engine][createIndex], indexName, t.Name, column.Name)
@@ -202,6 +220,7 @@ func (t *Table) createIndex(ctx context.Context, db *sql.DB, engine string, colu
 	return err
 }
 
+// initView initializes the view by checking if it exists and creating it if necessary.
 func (v *View) initView(db *sql.DB, engine string) error {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
@@ -215,6 +234,7 @@ func (v *View) initView(db *sql.DB, engine string) error {
 	return nil
 }
 
+// isExists checks if the view exists in the database.
 func (v *View) isExists(ctx context.Context, db *sql.DB, engine string) bool {
 	query := fmt.Sprintf(queries[engine][checkView], v.Name)
 	_, err := db.ExecContext(ctx, query)
@@ -222,6 +242,7 @@ func (v *View) isExists(ctx context.Context, db *sql.DB, engine string) bool {
 	return err == nil
 }
 
+// createView creates the view in the database.
 func (v *View) createView(ctx context.Context, db *sql.DB, engine string) error {
 	query := fmt.Sprintf(queries[engine][createView], v.Name, v.Query[engine])
 	_, err := db.ExecContext(ctx, query)